ATM Interface
We have all come across ATMs in our cities and it is built on Java. This complex project cc five different classes and is a console-based application. When the system starts the use prompted with user id and user pin. On entering the details successfully, then ATM funct are unlocked. The project allows to perform following operations:

1. Transactions History

2. Withdraw

3.Deposit

4. Transfer

5.Quit